backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aioredis==1.3.1
    # via
    #   -r requirements.in
    #   channels-redis
asgiref==3.8.1
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
    #   django-cors-headers
async-timeout==5.0.1
    # via aioredis
attrs==24.2.0
    # via
    #   service-identity
    #   twisted
autobahn==24.4.2
    # via daphne
automat==24.8.1
    # via twisted
cffi==1.17.1
    # via cryptography
channels==3.0.5
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==3.3.1
    # via -r requirements.in
constantly==23.10.4
    # via twisted
cryptography==43.0.3
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==3.0.2
    # via channels
django==5.0.1
    # via
    #   -r requirements.in
    #   channels
    #   django-cors-headers
    #   django-redis
    #   djangorestframework
django-cors-headers==4.6.0
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via -r requirements.in
hiredis==3.0.0
    # via aioredis
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.10
    # via
    #   hyperlink
    #   twisted
incremental==24.7.2
    # via twisted
msgpack==1.1.0
    # via channels-redis
pillow==11.0.0
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.1
    # via service-identity
pycparser==2.22
    # via cffi
pyopenssl==24.2.1
    # via twisted
redis==5.2.0
    # via django-redis
service-identity==24.2.0
    # via twisted
sqlparse==0.5.2
    # via django
twisted[tls]==24.10.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.12.2
    # via twisted
tzdata==2024.2
    # via django
unidecode==1.3.8
    # via -r requirements.in
zope-interface==7.1.1
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
